This project uses web scraping with Beautiful Soup and Selenium to do data entry automation.
I am BeautifulSoup to do web scraping on the Zillow website. The website compiles a list of apartments for rent in the Chicago area. The apartments are to have 1+ bedrooms and the cost should be under $3,000.

I generate a list of links, prices, and addresses for each listing. I then use Selenium to enter the data to a Google form which I can later convert to a nicely formatted Google Sheet document for viewing.